backend/requirements.txt
"setuptools>=68.0.0
fastapi==0.115.12
uvicorn==0.29.0
motor==3.3.2
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.1
pydantic==2.7.1
email-validator==2.1.1" | Out-File -Encoding utf8 requirements.txt